minor changes create a enums folder and move paths there
