+++ src/main/java/com/solvd/enums/Paths.java
@@ -1,4 +1,4 @@
-package com.solvd;
+package com.solvd.enums;
 
 public enum Paths {
